fixes for stress mock (#172)
* fixes for stress mock
