+++ examples/stress-mock/deployment/deployment.yaml
@@ -23,6 +23,12 @@ spec:
           env:
             - name: DEBUG
               value: "true"
+            - name: OPENAPI
+              value: "100"
+            - name: ODATA
+              value: "100"
+            - name: EVENT
+              value: "100"
 ---
 apiVersion: v1
 kind: Service
